app/src/main/java/dev/propoc/flickrcodechallenge/viewmodel/SearchViewModel.kt
package dev.propoc.flickrcodechallenge.viewmodel

import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import dev.propoc.flickrcodechallenge.model.ImageResponseModel
import dev.propoc.flickrcodechallenge.repository.Repository
import kotlinx.coroutines.flow.MutableStateFlow
import kotlinx.coroutines.flow.StateFlow
import kotlinx.coroutines.flow.asStateFlow
import kotlinx.coroutines.launch

class SearchViewModel : ViewModel() {
    private val repository = Repository()

    private val _dataList = MutableStateFlow<List<ImageResponseModel>>(emptyList())
    val dataList: StateFlow<List<ImageResponseModel>> = _dataList.asStateFlow()

    private val _isLoading = MutableStateFlow(false)
    val isLoading: StateFlow<Boolean> = _isLoading.asStateFlow()

    private val _errorMessage = MutableStateFlow<String?>(null)
    val errorMessage: StateFlow<String?> = _errorMessage.asStateFlow()

    fun fetchData(name: String) {
        viewModelScope.launch {
            _isLoading.value = true
            _errorMessage.value = null
            try {
                val data = repository.getData(name)
                _dataList.value = listOf(data)
            } catch (e: Exception) {
                _errorMessage.value = "Failed to load data: ${e.message}"
            } finally {
                _isLoading.value = false
            }
        }
    }
}